+++ ios/Classes/EdfaPaySdkEventChannels.swift
@@ -13,9 +13,10 @@ import UIKit
 
 public class EdfaPaySdkEventChannels: NSObject{
     
-    var cardpay:FlutterEventChannel? = nil;
-    var cardDetailPay:FlutterEventChannel? = nil;
-    var applepay:FlutterEventChannel? = nil;
+    var sadadpay:FlutterEventChannel? = nil
+    var cardDetailPay:FlutterEventChannel? = nil
+    var cardpay:FlutterEventChannel? = nil
+    var applepay:FlutterEventChannel? = nil
     
     var sale:FlutterEventChannel? = nil;
     var recurringSale:FlutterEventChannel? = nil;
@@ -32,6 +33,7 @@ public class EdfaPaySdkEventChannels: NSObject{
 
         cardpay = FlutterEventChannel(name: "com.edfapg.flutter.sdk.cardpay", binaryMessenger: messenger)
         cardDetailPay = FlutterEventChannel(name: "com.edfapg.flutter.sdk.cardDetailPay", binaryMessenger: messenger)
+        sadadpay = FlutterEventChannel(name: "com.edfapg.flutter.sdk.sadadpay", binaryMessenger: messenger)
         applepay = FlutterEventChannel(name: "com.edfapg.flutter.sdk.applepay", binaryMessenger: messenger)
         
         sale = FlutterEventChannel(name: "com.edfapg.flutter.sdk.sale", binaryMessenger: messenger)
@@ -46,6 +48,7 @@ public class EdfaPaySdkEventChannels: NSObject{
         
         cardpay?.setStreamHandler(CardPayEventHandler())
         cardDetailPay?.setStreamHandler(CardDetailPayEventHandler())
+        sadadpay?.setStreamHandler(SadadPayEventHandler())
         applepay?.setStreamHandler(ApplePayEventHandler())
         
         sale?.setStreamHandler(SaleEventHandler())